refactor(chat): extract message decoder out of ChatMessageStream
ChatMessageStream.kt was 841 lines, with the entire HomebaseFile →
MessageUiModel decoder (mapToMessageData + getDeliveryStatus +
renderStatusMessage) wrapped in a companion object — ~440 lines of pure
mapping with no need for the surrounding stream/lifecycle context.

Move the three functions verbatim into a new top-level MessageMapper.kt
in the same package. ChatMessageStream now holds only stream/lifecycle
concerns (event loop, processIncrementalBatch, observe/load/fetch/search
methods, conversationState cache, MessageLookup impl), down to 397 lines.

Also drops a stack of imports that moved with the decoder
(TranslationUtil, all MR.string.system_*, ReactionContent,
MessageContent, MessageContentParser, JsonPrimitive). External callers
in ConversationMapper.kt and ConversationStream.kt switch from
ChatMessageStream.mapToMessageData(...) to the same-package top-level
mapToMessageData(...) via a one-line import swap.

Pure mechanical move — no behavior change. The two existing decoder
tests (ChatMessageStreamMapperTest, ChatMessageStreamPagingTest) cover
the move; both updated to drop the ChatMessageStream. qualifier and
keep working unchanged otherwise. Adds a new
MessageMapperDeliveryStatusTest to lock in the 7 branches of
getDeliveryStatus, which had no direct unit coverage before.

Sets up the eventual paging PR — the decoder no longer balloons
ChatMessageStream when paging adds a MessageWindow / cursor-pair holder.
